Added more user attributes, incorporated them into the entire website.

+++ server/database/migrations/2021_03_11_200232_create_boat_user_table.php
@@ -18,7 +18,7 @@ class CreateBoatUserTable extends Migration
             $table->id();
             $table->unsignedBigInteger('boat_id');
             $table->unsignedBigInteger('user_id');
-            $table->integer('role')->default(BoatUser::ROLE_CREW);
+            $table->tinyInteger('role')->default(BoatUser::ROLE_CREW);
             $table->timestamps();
 
             $table->foreign('boat_id')
